pyproject.toml
[project]
name = "fastapiv"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "dotenv>=0.9.9",
    "fastapi[standard]",
    "psycopg[binary]>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "factory-boy>=3.3.3",
    "freezegun>=1.5.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "taskipy>=1.14.1",
]

[tool.pytest.ini_options]
pythonpath = "."

